refactor: move app window to windows folder
